[BE][MAIN] LMS-197 Pass test from menu
