+++ src/interfaces/test_result.model.ts
@@ -1,6 +1,6 @@
 export interface ITestResultModel {
   _id?: string;
-  lesson_id: string;
+  lesson_id?: string;
   result: number;
   questions_id: string[];
 }
